More changes, MemoryGame, download, etc

+++ Classes/Album.m
@@ -137,7 +137,7 @@
 }
 
 - (void)parser:(NSXMLParser *)parser parseErrorOccurred:(NSError *)parseError {
-	NSLog(@"Error Parsing at line: %i, column: %i", parser.lineNumber, parser.columnNumber);	
+	NSLog(@"Error Parsing at line: %li, column: %li", (long)parser.lineNumber, (long)parser.columnNumber);
 }
 
 -(void) dealloc {
