+++ Classes/LandscapeManager.m
@@ -45,7 +45,7 @@ int qtyLandscapes, currentLandscape = 0;
 }
 
 + (void)parser:(NSXMLParser *)parser parseErrorOccurred:(NSError *)parseError {
-	NSLog(@"LandscapeManager. Error Parsing at line: %i, column: %i", parser.lineNumber, parser.columnNumber);	
+	NSLog(@"LandscapeManager. Error Parsing at line: %li, column: %li", (long)parser.lineNumber, (long)parser.columnNumber);
 }
 
 + (Landscape *) getCurrentLandscape {
